LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/container-units-selection-expected.txt

Test

FAIL Container units select the proper container assert_equals: expected "30px" but got "10px"
FAIL Units respond to the writing-mode of the element assert_equals: expected "40px" but got "30px"
